serial: 梳理 read/write 返回值

+++ libserial/include/serial.h
@@ -1,9 +1,9 @@
 #ifndef __LIB_SERIAL__
 #define __LIB_SERIAL__
 
-#include <stdio.h>
-#include <stdint.h>
 #include <stdbool.h>
+#include <stdint.h>
+#include <stdio.h>
 
 struct _serial_dev_t;
 typedef struct _serial_dev_t serial_dev_t;
@@ -13,8 +13,8 @@ void serial_close(serial_dev_t **dev);
 
 bool serial_set_speed(serial_dev_t *dev, uint32_t speed);
 
-int32_t serial_read(serial_dev_t *dev, void *buf, size_t count, uint64_t timeout);
-int32_t serial_write(serial_dev_t *dev, const void *buf, size_t count, uint64_t timeout);
+ssize_t serial_read(serial_dev_t *dev, void *buf, size_t count, uint64_t timeout);
+ssize_t serial_write(serial_dev_t *dev, const void *buf, size_t count, uint64_t timeout);
 
 void serial_discard_input(serial_dev_t *dev);
 void serial_discard_output(serial_dev_t *dev);
